Refactoring
* Add enumerations

* Format, lint, etc.

* Add tests for cwl.py

* Remove unused lines

* Fix out of order step resolution

To create the needed steps the algorithm was finding the output
nodes and traversed upwards extended the to_find queue.

However there where two issues with that approach:

1. steps that share inputs will result in adding the same node twice
in the needed_steps list.

2. There are workflow outputs that can participate as inputs in
intermediate steps.

These changes make sure that already visited nodes are not added to
the needed_steps and that we start from outputs that are not intermediate
inputs.

* Fix

* Test

* Remove jobs for now

* Simple ci

* Make inputs outputs homogeneous

* Add cwl enumerations

* Support both list and dict io parameters

* Cleanup

* Allow passing File as cmd argument